Reorganized project structure (closes #3)
