--- SM_Server.csproj
@@ -1,32 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>SMServer</RootNamespace>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="Packets\BinEndianBinaryWriter.cs" />
-    <Compile Remove="Packets\Checksum.cs" />
-    <Compile Remove="Packets\ChecksumAccepted - Copy %282%29.cs" />
-    <Compile Remove="Packets\ChecksumAccepted - Copy.cs" />
-    <Compile Remove="Packets\helper.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="Packets\JoinConfirmation" />
-    <None Remove="Packets\NewFile1.txt" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Facepunch.Steamworks" Version="2.3.3" />
-    <PackageReference Include="K4os.Compression.LZ4.Streams" Version="1.3.5" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="System.Text.Json" Version="7.0.2" />
-  </ItemGroup>
-
-</Project>
