+++ ScrapServer.Networking/Client/ClientConnectingEventArgs.cs
@@ -1,4 +1,4 @@
-﻿namespace SMServer.src.Client;
+﻿namespace ScrapServer.Networking.Client;
 
 public class ClientConnectingEventArgs
 {
