+++ ScrapServer.Networking/ScrapServer.Networking.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="../ScrapServer.Utility/ScrapServer.Utility.csproj"/>
+	</ItemGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Facepunch.Steamworks" Version="2.3.3" />
+	</ItemGroup>
+
+</Project>
