+++ ScrapServer.Utility/ScrapServer.Utility.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="K4os.Compression.LZ4.Streams" Version="1.3.5"/>
+	</ItemGroup>
+
+</Project>
