+++ ScrapServer/ScrapServer.csproj
@@ -1,0 +1,24 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Facepunch.Steamworks" Version="2.3.3" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\ScrapServer.Networking\ScrapServer.Networking.csproj" />
+	  <ProjectReference Include="..\ScrapServer.Utility\ScrapServer.Utility.csproj" />
+	</ItemGroup>
+
+	<Target AfterTargets="Build" Name="CopySteamDeps">
+		<Copy SourceFiles="steam_api64.dll" DestinationFolder="$(OutDir)"/>
+		<Copy SourceFiles="steam_appid.txt" DestinationFolder="$(OutDir)"/>
+	</Target>
+
+</Project>
