+++ ScrapServer/steam_appid.txt
@@ -1,0 +1,1 @@
+387990
